% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{create_output}
\usage{
create_output(peakList_A, peakList_B, tau.s = 1e-08, output_path = getwd(),
  ...)
}
\arguments{
\item{tau.s}{permissive threshold for stringent enriched regions, all enriched peaks below this threshold, are considered stringent peaks}

\item{peaklist_A}{set of confirmed enriched regions for all Chip-seq replicates}

\item{peaklist_B}{set of discarded enriched regions for all Chip-seq replicates}

\item{outDir}{user can control where exported BED file goes}
}
\value{
BED file
}
\description{
create stringnet/weak ERs by permissive stringent threshold
}
\details{
Filter ERs by permissive stringency threshold

Chip-seq detects genome-wide DNA protein interation, returing enriched regions which associated with significance score.
In our methodological framework, we design permissive threshold for all enriched regions' score (A.K.A, peak signal's significance),
to classify each peak as stringent or weakly enriched by its p-value.Once getting set of confirmed, discarded enriched regions for all Chip-seq replicates,
it is intuitive to generate stringent, weak peak set accordingly. To help user gaining deeper insight and biological evaluation of analysis result,
graphical visualization of corresponding peak set also created by handy.
}
\author{
Julaiti Shayiding
}

