% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_ERs.R
\name{denoise_ERs}
\alias{denoise_ERs}
\title{Data cleaning (exclude noise from input sample)}
\usage{
denoise_ERs(peakGRs, tau.w = 1e-04, nmtab = "noise", dest.dir = tempdir(),
  overwrite = FALSE)
}
\arguments{
\item{peakGRs}{Chip-seq replicate imported and
all enriched regions stored in \code{GRanges} objects}

\item{tau.w}{permissive p-value threshold for weakly
enriched peak, all enrichred regions' pvalue above
this thrshold, are considered background noise}

\item{nmtab}{Character string that assigned as name to the noise}

\item{dest.dir}{dirctory that exported files can be placed.
only noise can be exported as standard BED file;}

\item{overwrite}{default FALSE indicating whether
existing files with identical name should be over-written.}
}
\value{
Peaks without background noise is
stoted in \link[GenomicRanges]{GRangesList}
}
\description{
Given the output of \link{readPeakFiles}, We set up
permissive p-value threshold \code{tau.w} for weakly
enriched regions associated with score and p-value.
We are interested in moderately enriched regions
(A.K.A, weak peak), but extremely weakly enriched regions
won’t give any biological insight and must be removed from
all candidate sample. Note that p-value threshold can be
tuned by users, so using different threshold value will
result in different filtering results. Background signal
(A.K.A, noise) also exported as standard BED file by
using \link[rtracklayer]{export.bed} for the sake of
evaluate each ChIP-Seq replicate that bearing different
output set with clear biological evidence.
}
\details{
Passed to \link{runMSPC}
}
\examples{
require(rtracklayer)
require(GenomicRanges)

## read peak file as GRanges object
files <- getPeakFile()[1:3]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        nmtab = "noise", dest.dir = getwd(), overwrite = TRUE)
## Explore all stringent and weak enriched regions
total.ERs
}
\author{
Jurat Shahidin
}
