% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_stats.R
\name{FDR_stats}
\alias{FDR_stats}
\title{multiple testing correction}
\usage{
FDR_stats(peakList, pAdjustMethod = "BH", fdr = 0.05, asPlot = TRUE)
}
\arguments{
\item{peakList}{set of confirmed peaks through combined stringency test.}

\item{pAdjustMethod}{pvalue adjustment method}

\item{fdr}{parameter for false discovery rate}

\item{asPlot}{logical whether produce graphical plot or not}
}
\value{
set of enriched regions in BED format file
}
\description{
multiple testing correction
}
\author{
Jurat Shahidin
}
