% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByFisherMethod.R
\name{filterByFisherMethod}
\alias{filterByFisherMethod}
\title{filterByFisherMethod}
\usage{
filterByFisherMethod(peakset, .hitList, cmbstrgThreshold = 1e-08,
  isFisherPass = c(TRUE, FALSE), ...)
}
\arguments{
\item{peakset}{set of pre-processed enriched regions stored in GRanges, where background signal (A.K.A, noise peak) were pre-processed and won't involve further workflow}

\item{cmbstrgThreshold}{combined stringency threshold against all enriched regions p-value, and each peaks are classified as confirmed or discarded peak accordingly.}

\item{isFisherPass}{logical vector that check whether all enriched regins are passed in Fisher' combined method. TRUE : return GRangesList of all enriched regions are classified as confirmed. FALSE: return GRangesList of all discarded peaks. both is required}

\item{.hitLiist}{list of overlap hit that set of enriched regions comply minimum overlapping peak requirement.}
}
\description{
perform combined stringency test
}
\details{
Filter ERs by combined stringency test & create output set

Main idea behind our method is that repeated evidence across Chip-seq replicates can compensate for a
lower significance in a single sample, which is implemented though the Fisher method. The significace of
ovelapping enriched regions is rigorously combined with the Fisher's method to obtain global Fisher score (A.K.A, combined p-value)
to evaluate combined stringency of all enriched regions and classified each peak as confirmed or discarded.
}
\author{
Julaiti Shayiding
}

