% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByOverlapHit.R
\name{filterByOverlapHit}
\alias{filterByOverlapHit}
\title{filterByOverlapHit}
\usage{
filterByOverlapHit(.ovHit, peakset, replicate.type = c("Biological",
  "Technical"), isSuffOverlap = TRUE)
}
\arguments{
\item{.ovHit}{list of overlap hit index. Once we obtained overlap hit list by calling `peakOverlapping` function in previous workflow, resuled hit-list ready to be used.}

\item{peakset}{set of Chip-seq replicate imported and all peaks are stored in GRanges object.}

\item{replicate.type}{A charcter vector used to select type of Chip-seq replicate ( Biological / Technical replicate)}

\item{isSuffOverlap}{logical vector that check whether comply minimum overlapping peak requirement. TRUE: return overlap hit list where all enriched regions comply minimum overlapping peak requirement. FALSE : return all discarded enriched regions as GRange object due to failing to comply minimum overlapping peak requirement. both is required}
}
\value{
list-like object
}
\description{
Filter ERs by overlap-hit index list
}
\details{
Once we obtained list of overlap-hit index where all possible overlap pair is included, we need to
further investigate hit index with our defined method. In previous workflow, `peakOverlapping` function give us correct geometry of overlap peaks as list-like object by parallel.
we need to do vector sum for getting overall overlaped peak numbers and proceed our evaluation. Where all enriched regions comply minimum overlap peak requirement,
are returned as overlap hit list and used for next workflow, while all enriched regions fail to comply the minimum overlap peak requirement, are returned as GRanges objects.
Due to further need of evaluating all enriched regions in different level in order to give clear Biological evidence, we decide to expand those overlap hit index as GRanges objects.
}
\author{
Julaiti Shayiding
}

