% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_stats.R
\name{Fisher_stats}
\alias{Fisher_stats}
\title{Retrieve Fisher's global score through Fisher Method}
\usage{
Fisher_stats(hitList, peakset)
}
\arguments{
\item{hitList}{output of \link{filterByOverlapHit} take \code{isSuffOverlap}
as \code{TRUE}. Only Enriched regions that comply with
minimum overlapping peak requirement can be further evaluated. We kept ERs
that fulfill minimum overlapping requirement in \link[IRanges]{IntegerList}.}

\item{peakset}{output of \link{denoise_ERs}. Set of Chip-seq replicate
imported and all peaks are stored in GRanges object, where all
background noise won't be further processed.}
}
\value{
numeric vector, Global Fisher's score are in vector
}
\description{
We assess the presence of overlapping enrichred regions across
multiple Chip-seq replicates. Therefore, the significance of
overlapping enriched regions are rigorously combined
with Fisher's method to obtain global Fisher score.
Using \link{fisherCmbp} to get global Fisher'score.
}
\details{
Passed to \link{filterBycombStringency}
}
\examples{
library(GenomicRanges)
library(rtracklayer)

# read peak file as GRanges object
files <- getPeakFile()[7:8]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        fileName = "noise", outDir = getwd())
## find peak overlapping
hit <- peakOverlapping(total.ERs, FUN=which.max)

## check whether each ERs comply with minimum overlapping peak requirement
keepList <- filterByOverlapHit(hit, peakset = total.ERs,
                               replicate.type = "Biological",
                               isSuffOverlap=TRUE)

## Global Fisher' score
comb.p <- Fisher_stats(hitList = keepList , peakset = total.ERs)

}
\author{
Julaiti Shayiding
}

