% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fisherCmbp}
\alias{fisherCmbp}
\title{helper function to get combine p-value}
\usage{
fisherCmbp(p)
}
\arguments{
\item{p}{pvalue vector}
}
\value{
combined pvalue
}
\description{
helper function to get combine p-value
}
\examples{
## pvalue vector
df <- data.frame(v1=c(1e-02,1e-11, 1e-09),
                   v2=c(1e-18,1e-07, 1e-13))
## combined pvalue
cmbP <- apply(df,1,fisherCmbp)

}
\author{
Julaiti
}

