% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.pvalue}
\alias{get.pvalue}
\title{retrieve pvalue of ERs that comply minimum overlapping peak requirement}
\usage{
get.pvalue(hit, gr)
}
\arguments{
\item{hit}{overlap hit}

\item{gr}{GRanges object}
}
\value{
pvalue
}
\description{
retrieve pvalue of ERs that comply minimum overlapping peak requirement
}
\examples{
require(GenomicRanges)
require(XVector)

## example peak interval
grs <- GRangesList(
    bar = GRanges(
        seqnames=Rle("chr1", 3),ranges=IRanges(c(6,25,40), c(12,33,49)),
        strand = Rle(c("*"),3), rangeName=c("a2","a5","a8"),
        score=c(3,12,4),p.value=c(1e-03,1e-12,1e-04)),
    cat = GRanges(
        seqnames=Rle("chr1", 3),ranges=IRanges(c(15,19,47), c(18,28,55)),
        strand = Rle(c("*"),3), rangeName=c("b4","b6","b9"),
        score=c(11,3,6),p.value=c(1e-11,1e-03,1e-06))
)

## find overlap
hit <- peakOverlapping(grs, which.max)

## explore pvalue
pv <- Map(get.pvalue, hit, grs)
}
\author{
Julaiti Shayiding
}

