% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlot.R
\name{getPlot}
\alias{getPlot}
\title{Graphical view of different ERs set for each Chip-seq replicates.}
\usage{
getPlot(peakList_A, peakList_B, tau.s = 1e-08)
}
\arguments{
\item{peakList_A}{output of \link{runMSPC},
is set of all confirmed ERs in \link[GenomicRanges]{GRanges} objects.}

\item{peakList_B}{output of \link{runMSPC},
is set of all discarded ERs in \link[GenomicRanges]{GRanges} objects.}

\item{tau.s}{permissive threshold for stringent enriched regions,
all enriched regions below this threshold, are considered stringent ERs}
}
\value{
using \link[ggplot2]{ggplot} to generate stack bar plot for file bar
}
\description{
This function is served as graphical version of \link{export_ERs}.
To help user gaining deeper insight and biological evaluation of
analysis result, using \link[ggplot2]{ggplot} to generate
stack bar plot for each Chip-seq replicates can be done.
}
\examples{
# set up
library(GenomicRanges)
library(rtracklayer)

# load peak files
files <- getPeakFile()[1:3]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        overwrite = TRUE)

## explore set of confirmed, discarde peaks
confirmedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = TRUE)
discardedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = FALSE)

# Visualize the output set for file bar
getPlot(peakList_A = confirmedERs,
        peakList_B = discardedERs, tau.s = 1.0E-08)
}
\author{
Jurat Shahidin
}
