% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPC_package.R
\docType{package}
\name{MSPC-package}
\alias{Calling}
\alias{MSPC-package}
\alias{Multiple}
\alias{Peak}
\alias{Sample}
\title{Bioconductor Package for Multiple Sample Peak Calling}
\description{
This package is designed for chip-seq data analysis
}
\details{
\tabular{ll}{ Package: \tab MSPC\cr Type: \tab Package
\cr Version: \tab 0.99.0\cr Date: \tab 15-11-2016
\cr biocViews:\tab MSPC, Annotation, Software
\cr Depends:\tab \cr Imports:
\tab methods, extractList, findOverlaps
\cr License: \tab GPL-3.0\cr }
}
\author{
Julaiti Shayiding

Maintainer: Julaiti Shayiding \email{juratbupt@gmail.com}
}
\keyword{package}

