% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R, R/mspc.R, R/read_inputs.R
\name{get_files}
\alias{get_files}
\alias{mspc}
\alias{read_inputs}
\title{Runs the mspc program on R}
\usage{
get_files(path)

mspc(
  Input,
  Input_folder = NULL,
  Directory_Granges_Inputs = NULL,
  Mspc_path = NULL,
  Replicate_Type,
  Stringency_threshold,
  Weak_threshold,
  Gamma = NULL,
  C = NULL,
  Alpha = NULL,
  Multiple_Intersections = NULL,
  Degree_of_Parallelism = NULL,
  Input_Parser_Configuration = NULL,
  Output_path = NULL,
  Granges = FALSE,
  Keep = NULL
)

read_inputs(Input, Directory_Granges_Inputs)
}
\arguments{
\item{path}{A directory path}

\item{Input}{Name of the inputs given ( Bed files or Granges objects). It could be a one sample file or multiple sample files. More information about the type of the inputs mspc can handle in Details}

\item{Input_folder}{Path of the location of the input files.}

\item{Directory_Granges_Inputs}{Folder path : when the inputs given are GRanges objects, bed files are created from these Granges objects. Where should they be stored ? The default value of the argument is the current working directory.}

\item{Mspc_path}{Installation path of the mspc executable.}

\item{Replicate_Type}{Character string : possible values : "Bio", "Biological", "Tec", "Technical". This argument defines the replicate type.}

\item{Stringency_threshold}{Integer : a threshold on p-values, where peaks with p-value lower than this threshold, are considered stringent.}

\item{Weak_threshold}{Integer : a threshold on p-values, such that peaks with p-value between this and stringency threshold, are considered weak peaks.}

\item{Gamma}{Integer : the combined stringency threshold. Peaks with combined p-value below this threshold are confirmed.}

\item{C}{Integer : the minimum number of overlapping peaks required before MSPC combines their p-value.}

\item{Alpha}{Integer : the threshold for Benjamini-Hochberg multiple testing correction.}

\item{Multiple_Intersections}{Character string : either "lowest" or "highest". When multiple peaks from a sample overlap with a given peak, this argument defines which of the peaks to be considered: the one with lowest p-value, or the one with highest p-value?}

\item{Degree_of_Parallelism}{Integer : the number of parallel threads MSPC can utilize simultaneously when processing data.}

\item{Input_Parser_Configuration}{Path to a JSON file containing the configuration for the input BED file parser.}

\item{Output_path}{Path in which analysis results should be persisted.}

\item{Granges}{Logical : determines whether or not the wrapper function should read the results of MSPC and store them as GRanges objects in the R environment.The default value is FALSE.}

\item{Keep}{Logical : determines whether the mspc function should keep or delete all the files generated while running the mspc program.If set to FALSE, mspc will automatically set the argument Granges to TRUE. More information on the default value of the argument in Details.}
}
\value{
List of files in a folder and subfolders.

A list of files, according to the values of the arguments. More information in Details

Name of input bed files. More information in Details
}
\description{
MSPC comparatively evaluates ChIP-seq peaks and combines the statistical significance
of repeated evidences, with the aim of lowering the minimum significance required
to "rescue" weak peaks; hence reducing false-negatives.
}
\details{
Only one type of arguments is supported by mspc at a time, ie, the user can either
give all the inputs as bed files, or all the inputs as Granges objects.
Therefore, the user cannot give an input argument that contains
bed files and GRanges objects.
If the user gives the inputs as Granges objects, he can either give a single Granges object
or a list of Granges objects.
Similarly, the user can either give one bed file or a character string of multiplie bed file names.

The mspc function prints the summary of the mspc program results.

When the user gives inputs as a bed file, the function returns the name of these bed files.
When the user gives inputs as Granges object, it exports it into bed files and returns
the name(s) of the bed files created by the function.
The bed files are created in the directory specified by Directory_Granges_Inputs
The names of the bed files created will be given to the mspc function as inputs.
}
\section{Functions}{
\itemize{
\item \code{get_files}: Get a list of files in a folder

\item \code{read_inputs}: Reads the inputs given by the user to the mspc function
}}
