% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakClassifier.R
\name{peakClassifier}
\alias{peakClassifier}
\title{Identify stringent/weak enriched regions by permissive threshold}
\usage{
peakClassifier(peakList_A, peakList_B, tau.s = 0.00000001,
  graphical.output = TRUE, exportType = c("bed", "csv"))
}
\arguments{
\item{peakList_A}{set of confirmed peaks}

\item{peakList_B}{set of discarded peaks}

\item{tau.s}{permissive threshold for stringent peaks}

\item{graphical.output}{logical vector}
}
\value{
set of enriched regions in data.frame
}
\description{
Identify stringent/weak enriched regions by permissive threshold
}
\author{
Jurat Shahidin
}
