% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeakFiles.R
\name{readPeakFiles}
\alias{readPeakFiles}
\title{readPeakFile}
\usage{
readPeakFiles(peakFolder, pvalueBase = 1L, verbose = FALSE, ...)
}
\arguments{
\item{peakFolder}{set of ChipSeq peak replicate in BED format}

\item{pvalueBase}{user has options to choose the p-value format(- log(p-value), -10 log(p-value), -100 log(p-value)). By default, use -10 log(p-value) ;}

\item{verbose}{logical that control whether the output is printed or not}
}
\value{
GRanges object
}
\description{
read Chip-seq replicated and all enriched regiions are stored in GRanges object.
}
\details{
Import Chip-seq replicates and all ERs are stored in GRanges objects.

readPeakFiles can read peak file in standard BED format and stored in GRanges object,
where several peak files (A.K.A, Chip-seq replicates) can be read simultaneously using lapply
}
\author{
Julaiti Shayiding
}

