% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSPC.R
\name{runMSPC}
\alias{runMSPC}
\title{Assess overlapping enriched regions across multiple sample}
\usage{
runMSPC(peakset, whichType = c("max", "min"),
  replicate.type = c("Biological", "Technical"), cmbStrgThreshold = 1e-08,
  isConfirmed = TRUE)
}
\arguments{
\item{peakset}{the output of \link{denoise_ERs}.
set of Chip-seq replicate imported and all peaks are stored in
GRanges object, where all background signal were excluded.}

\item{whichType}{user has options to keep most stringent(with loest p-value)
or least stringent(highest p-value) peak if multiple overlapping peaks
were detected. By default, keep most stringent one.}

\item{replicate.type}{A charcter vector used to select type of input
Chip-seq replicate ( Biological / Technical replicate).}

\item{cmbStrgThreshold}{combined stringency threshold
against all enriched regions p-value, so we could identify
whether ERs fulfill combined stringency test, and result can be
set of confirmed/discarded peaks respectively.}

\item{isConfirmed}{logical vector that check whether ERs
comply combined stringency test.}
}
\value{
\code{isConfirmed} is \code{True}:
return set of peaks that  both passed from minimum overlapping
peak criteria and combined stringency test (Fisher method)

\code{isConfirmed} is \code{False} :
return set of peaks either  failed from Fisher method or
minimum overlapping requirement.
}
\description{
We assess the presence of overlapping enriched regions
(A.K.A peaks) across multiple sample simultaneously,
repeated evidence over multiple replicates can compensate
for lower significance in a single sample, using Fisher method
to increase the statistical significance of weak evidence which
might get involved in TF or gene regulation.
}
\note{
Special thanks to Martin Morgan's contribution on this revision
}
\examples{
# set up
library(GenomicRanges)
library(rtracklayer)

# load peak files
files <- getPeakFile()[1:3]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        overwrite = TRUE)

## explore set of confirmed, discarde peaks
confirmedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = TRUE)
discardedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = FALSE)
}
\author{
Jurat Shahidin
}
