% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Graphical view of different output for each Chip-seq replicates.}
\usage{
create_output(peakList_A, peakList_B, tau.s = 1e-08)
}
\arguments{
\item{peakList_A}{output of \link{filterBycombStringency},
is set of all confirmed ERs in \link[GenomicRanges]{GRanges} objects.}

\item{peakList_B}{output of \link{filterBycombStringency},
is set of all discarded ERs in \link[GenomicRanges]{GRanges} objects.}

\item{tau.s}{permissive threshold for stringent enriched regions,
all enriched regions below this threshold, are considered stringent ERs}
}
\value{
using \link[ggplot2]{ggplot} to generate stack bar plot for file bar
}
\description{
Chip-seq detects genome-wide DNA protein interation,
returing enriched regions which associated with significance score.
Using permissive threshold tau.s for signal' significant
value of stringent enriched region, we could identify
set of stringent, weakly enriched regions by manipulating the output of
\link{filterBycombStringency} function.  To help user gaining
deeper insight and biological evaluation of analysis result, using \link[ggplot2]{ggplot}
to generate stack bar plot for each Chip-seq replicates can be done.
}
\author{
Julaiti Shayiding
}

