% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_ERs.R
\name{denoise_ERs}
\alias{denoise_ERs}
\title{denoise_ERs}
\usage{
denoise_ERs(peakGRs, tau.w = 1e-04, .fileName = "", outDir = getwd(),
  verbose = FALSE, ...)
}
\arguments{
\item{peakGRs}{list of Chip-seq replicate imported and all enriched regions stored in GRanges objects}

\item{tau.w}{threshold value for weakly enriched peak, all enrichred regions' pvalue above this thrshold, are considered background signal (A.K.A, noise peak)}

\item{.fileName}{user has option to name background signal by their own preference.}

\item{outDir}{user control where the exported BED files goes}

\item{verbose}{control whether the output is printed or not}
}
\value{
GRangesList
}
\description{
pre-processing set of Chip-seq replicate & exclude all noise peak from samples
}
\details{
we set up threshold for signal' significant value of each enriched region,
where extremely weakly enriched regions won't be processed, so we did purification on original input dataset;
Extremenly weakly enriched regions (A.K.A, background noise signals) are exported as standard BED file for the sake of giving clear biological evidence.
The purpose of exposting noise peak as BED file, we'll evaluate each Chip-seq replicates that bearing different output set with clear biological evidence.
}
\author{
Julaiti Shayiding
}

