% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_ERs.R
\name{denoise_ERs}
\alias{denoise_ERs}
\title{Exclude all background signal (a.k.a, noise) from input Chip-seq samples}
\usage{
denoise_ERs(peakGRs = NULL, tau.w = 1e-04, fileName = "noise",
  outDir = getwd(), verbose = FALSE)
}
\arguments{
\item{peakGRs}{list of Chip-seq replicate imported and
all enriched regions stored in \code{GRanges} objects}

\item{tau.w}{threshold value for weakly enriched peak,
all enrichred regions' pvalue above this thrshold,
are considered background signal (a.k.a, noise)}

\item{fileName}{user has option to name background signal
by their own preference. by default name it as "noise"}

\item{outDir}{only noise can be exported as standard BED file;
user has an option to control where the exported files goes}

\item{verbose}{control whether the output is printed or not}
}
\value{
list of enriched regions stored in \link[GenomicRanges]{GRanges}
}
\description{
Give the output of \link{readPeakFiles}, this function
exclude all bacckground noise from input Chip-seq replicates.
using permissive threshold tau.w for signal' significant
value of each enriched region. Extremely weakly enriched regions
won't be processed and excluded from input Chip-seq replicates.
Background signal (a.k.a, noise) also exported as standard BED file
by using \link[rtracklayer]{export.bed} for the sake of
evaluate each Chip-seq replicate that bearing
different output set with clear biological evidence.
}
\details{
Passed to \link{peakOverlapping}
}
\examples{
require(rtracklayer)
require(GenomicRanges)

## read peak file as GRanges object
files <- getPeakFile()[7:8]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        fileName = "noise", outDir = getwd())
## Explore all stringent and weak enriched regions
total.ERs
}
\author{
Julaiti Shayiding
}
