% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ERs.R
\name{export_ERs}
\alias{export_ERs}
\title{Identify and Export stringent/ weak ERs}
\usage{
export_ERs(peakList_A, peakList_B, tau.s = 1e-08, exportFormat = c("bed",
  "csv"))
}
\arguments{
\item{peakList_A}{output of \link{runMSPC}, is set of ERs
that fulfill combined stringency test, rescued by Fisher's method,
also known as confirmed ERs.}

\item{peakList_B}{output of \link{runMSPC}, is set of discarded ERs
that failing from combined stringency test and minimum overlapping peak requirement}

\item{tau.s}{permissive threshold value for stringent ERs,
all ERs' pvalue below this value, are considered as stringent ERs.}

\item{exportFormat}{user has an option to export the result either BED format file or csv.}
}
\value{
stringent/weak peak set
}
\description{
ChIP-Seq detects genome-wide DNA protein interaction,
returning enriched regions which associated with significance score.
Using permissive threshold tau.s for stringent enriched region,
we could identify set of stringent, weakly enriched regions by the
output of \link{runMSPC}. All ERs in different group can be
exported as either BED format file by using \link[rtracklayer]{export.bed},
or csv by \link[utils]{write.csv}.
}
\examples{
# set up
library(GenomicRanges)
library(rtracklayer)

# load peak files
files <- getPeakFile()[1:3]
grs <- readPeakFiles(files, pvalueBase=1L)

## Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                        overwrite = TRUE)

## explore set of confirmed, discarde peaks
confirmedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = TRUE)
discardedERs <- runMSPC(peakset = total.ERs, whichType = "max",
                        cmbStrgThreshold = 1.0E-08, isConfirmed = FALSE)

## explore the peaks in different output set for file bar
export_ERs(peakList_A = confirmedERs,
           peakList_B = discardedERs,
           tau.s = 1.0E-08, exportFormat = "csv")

}
\author{
Jurat Shahidin
}
