% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_stats.R
\name{Fisher_stats}
\alias{Fisher_stats}
\title{Fisher_stats}
\usage{
Fisher_stats(.hitList, peakset, verbose = FALSE, ...)
}
\arguments{
\item{.hitList}{overlap hit index for all enrichred regions that comply minimum overlapping peak requirement.}

\item{peakset}{set of Chip-seq replicate imported and all peaks are stored in GRanges object, where all background noise pre-processed and won't involve in further downstream analysis.}
}
\value{
numeric vector
}
\description{
retrieve pvalue of ERs that comply minimum overlapping peak requirement and perfrom Fisher method
}
\details{
Combine stringecy test to rigorously combined overlapping peaks

we assess the presence of overlapping enrichred regions across multiple Chip-seq replicates.
Therefore, the significance of overlapping regions is rigorously combined with Fisher's method to obtain global Fisher score.
However, in next workflow, we are using Fisher combined p-value against combined stringency threshold
to evaluate combined stringency to all enriched regions in second level classification for all enriched regions.
We set up combined stringency threshold for all enriched regions and getting confirmed / discarded peak set accordingly.
}
\author{
Julaiti Shayiding
}

