% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_stats.R
\name{Fisher_stats}
\alias{Fisher_stats}
\title{Retrieve Fisher's global score through Fisher Method}
\usage{
Fisher_stats(hitList, peakset)
}
\arguments{
\item{hitList}{output of \link{filterByOverlapHit} take \code{isSuffOverlap}
as \code{TRUE}. Only Enriched regions that comply with
minimum overlapping peak requirement can be further evaluated. We kept ERs
that fulfill minimum overlapping requirement in \link[IRanges]{IntegerList}.}

\item{peakset}{output of \link{denoise_ERs}. Set of Chip-seq replicate
imported and all peaks are stored in GRanges object, where all
background noise won't be further processed.}
}
\value{
numeric vector, Global Fisher's score are in vector
}
\description{
We assess the presence of overlapping enrichred regions across
multiple Chip-seq replicates. Therefore, the significance of
overlapping enriched regions are rigorously combined
with Fisher's method to obtain global Fisher score.
Using \link[metap]{sumlog} to get global Fisher'score.
}
\details{
Passed to \link{filterBycombStringency}
}
\examples{
library(GenomicRanges)
library(rtracklayer)
library(XVector)
library(metap)

## Example Peaks in GRanges
bar=GRanges(
    seqnames=Rle("chr1", 3),ranges=IRanges(c(12,21,37), c(14,29,45)),
    strand=Rle(c("*"),3), rangeName=c("a1", "a2", "a3"), score=c(22, 6,13)
)

cat=GRanges(
    seqnames=Rle("chr1", 6),ranges=IRanges(c(5,12,16,21,37,78),
    c(9,14,19,29,45,84)),
    strand=Rle(c("*"),6), rangeName=c("b1", "b2","b3", "b4", "b6", "b7"),
    score=c(12, 5, 11, 8, 4, 3)
)
##Add p.value as metadata column
grs <- GRangesList("bar"=bar, "cat"=cat)
grs <- lapply(grs, pvalueConversion)

##Exclude background noise
total.ERs <- denoise_ERs(peakGRs = grs, tau.w = 1.0E-04,
                         fileName = "noise", outDir = getwd())

##peak overlapping
hit <- peakOverlapping(total.ERs, FUN=which.max)

##Retrieve pvalue of ERs that comply minimum overlapping peak requirement
keepList <- filterByOverlapHit(hit, peakset = total.ERs,
                               replicate.type = "Biological",
                               isSuffOverlap=TRUE)
\dontrun{
## Global Fisher' score
comb.p <- Fisher_stats(hitList = keepList , peakset = total.ERs)
comb.p
}
}
\author{
Julaiti Shayiding
}

