% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspc-package.R
\docType{package}
\name{mspc-package}
\alias{mspc}
\alias{mspc-package}
\title{mspc: MSPC program}
\description{
Runs the mspc program. MSPC comparatively evaluates ChIP-seq peaks and combines the statistical significance of repeated evidences, with the aim of lowering the minimum significance required to “rescue” weak peaks; hence reducing false-negatives.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
