% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakOverlapping.R
\name{peakOverlapping}
\alias{peakOverlapping}
\title{peakOverlapping}
\usage{
peakOverlapping(peakset, FUN = which.max)
}
\arguments{
\item{peakset}{set of Chip-seq replicate imported and all peaks are stored in GRanges object, where all background signal were pre-processed and won't involve in further downstream analysis.}

\item{FUN}{user has options to keep most stringent(with loest p-value) or least stringent(highest p-value) peak if multiple overlapping were detected. By default, keep most stringent overlaped peak.}
}
\value{
list-like object
}
\description{
asess peak overlapping across multiple sample and capture correct geometry of overlaping ERs
}
\details{
Asess peak overlapping across multiple Chip-seq replicate

we assess the presense of overlapping enriched regions across multiple Chip-seq replicates.
Each enriched regions in current replicates (A.K.A, chosen replicates) is evaluated
with the support of rest of Chip-seq replicates in input dataset for overlapping enriched regions by element-wise.
Due to processing each genomic region (all peak intervals are stored in GRanges objects) by element-wise to find overlap is quite inefficient,
peakOverlapping function efficiently vectorize overlapping peaks and retrieving overlapped regions from multiple Chip-seq replicates as list-like vector
where overlap position hit index gives us correct geomety of peak overlapping. It may happen that an enriched region(A.K.A, peak or ER) from current Chip-seq replicate overlap multiple ERs from other supported replicates.
peakOverlapping function retrieve only one overlapping peak from each supported Chip-seq replicate when multiple overlapping regions were detected.
}
\author{
Julaiti Shayiding
}

