% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeakFiles.R
\name{readPeakFiles}
\alias{readPeakFiles}
\title{Import Chip-seq replicates and all ERs are stored in GRanges objects.}
\usage{
readPeakFiles(peakFolder, pvalueBase = 10L, verbose = FALSE)
}
\arguments{
\item{peakFolder}{set of ChipSeq peak replicate in BED format file}

\item{pvalueBase}{user has options to choose the p-value format}

\item{verbose}{logical that control whether the output is printed or not}
}
\value{
GRanges object
}
\description{
readPeakFiles can read peak file in standard BED format using \link[rtracklayer]{import.bed}
and stored in GRanges object, where several peak files can be read simultaneously using lapply
}
\details{
Passed to \link{denoise_ERs}
}
\author{
Julaiti Shayiding
}

