% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeakFiles.R
\name{readPeakFiles}
\alias{readPeakFiles}
\title{Import Chip-seq replicates and all ERs are stored in GRanges objects.}
\usage{
readPeakFiles(peakFolder, pvalueBase = 1L, verbose = FALSE)
}
\arguments{
\item{peakFolder}{set of ChipSeq peak replicate in BED format file}

\item{pvalueBase}{User has option to select
p-value format (- log(p.value), -10 log(p.value), -100 log(p.value))}

\item{verbose}{logical that control whether the output is printed or not}
}
\value{
GRanges object
}
\description{
readPeakFiles can read peak file in standard BED format
using \link[rtracklayer]{import.bed} and stored in
\link[GenomicRanges]{GRanges} object, where several peak files
can be read simultaneously using lapply. Note that some data
sources provides Chip-seq enriched regions (A.K.A, peaks)
without p-value under specific conditions.
}
\details{
Passed to \link{denoise_ERs}
}
\examples{
require(rtracklayer)
require(GenomicRanges)
## get bed file
files <- getPeakFile()[1:3]
rd <- readPeakFiles(files, pvalueBase=1L)
}
\author{
Julaiti Shayiding
}

