% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeakFiles.R
\name{readPeakFiles}
\alias{readPeakFiles}
\title{Import Chip-seq replicates and all ERs are stored in GRanges objects.}
\usage{
readPeakFiles(peakFolder, pvalueBase = 1L, verbose = FALSE)
}
\arguments{
\item{peakFolder}{ChipSeq peak replicate either in BED format or data.frame}

\item{pvalueBase}{User has option to select
p-value format (- log(p.value), -10 log(p.value), -100 log(p.value))}

\item{verbose}{logical that control whether the output is printed or not}
}
\value{
GRanges object
}
\description{
This function can read peak file in standard BED file
using \link[rtracklayer]{import.bed} or data.frame like object
using \link[GenomicRanges]{makeGRangesFromDataFrame}
and stored in \link[GenomicRanges]{GRanges} object,
where several peak files can be read simultaneously
using lapply. Note that some data sources provide ChIP-Seq
enriched regions without p-value under specific
conditions. In our workflow, we are going to evaluate each
peak interval with p-value threshold, so represent peaks’ score
as p-value is needed.
}
\details{
Passed to \link{denoise_ERs}
}
\examples{
require(rtracklayer)
require(GenomicRanges)

## get bed file
files <- getPeakFile()[1:3]
readPeakFiles(files, pvalueBase=1L)
}
\author{
Jurat Shahidin
}
