% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSPC.R
\name{runMSPC}
\alias{runMSPC}
\title{Assess overlapping enriched regions across multiple sample}
\usage{
runMSPC(peakset, whichType = c("max", "min"),
  replicate.type = c("Biological", "Technical"),
  cmbStrgThreshold = 0.00000001, isConfirmed = TRUE)
}
\arguments{
\item{peakset}{the output of \link{denoise_ERs}.
set of Chip-seq replicate imported and all peaks are stored in
GRanges object, where all background signal were excluded.}

\item{whichType}{user has options to keep most stringent(with loest p-value)
or least stringent(highest p-value) peak if multiple overlapping peaks
were detected. By default, keep most stringent one.}

\item{replicate.type}{A charcter vector used to select type of input
Chip-seq replicate ( Biological / Technical replicate).}

\item{cmbStrgThreshold}{combined stringency threshold
against all enriched regions p-value, so we could identify
whether ERs fulfill stringency test,
and result can be set of confirmed/discarded peaks respectively.}

\item{isConfirmed}{logical vector that check whether ERs
comply combined stringency test.}
}
\value{
\code{isConfirmed} is \code{True}:
return set of enriched regions passed from combined stringency test

\code{isConfirmed} is \code{False} :
return set of enriched regions that failed from Fisher method
}
\description{
Assess overlapping enriched regions across multiple sample
}
\note{
Special thanks to Martin Morgan's contribution on this revision
}
\author{
Jurat Shahidin
}
